requirements.txt
# Python >= 3.11 (必须!)
requests>=2.25.1
html2text>=2020.1.16 
browser-use[memory]>=0.1.47
beautifulsoup4>=4.12.0
aiofiles>=23.2.1

# LLM提供商依赖 (根据需要安装)
langchain-openai>=0.0.5
#langchain-anthropic>=0.1.1
#langchain-google-genai>=0.0.5
langchain-deepseek>=0.1.1
#langchain-groq>=0.1.1

# 网络依赖
httpx[socks]>=0.25.0

python-dotenv>=1.0.0
playwright>=1.40.0 